New paint handling mechanism... uses BitSets from .gif (built-in brushes) or .png (user-created brushes).  button on toolbar for Paint Brush Editor with load/save cabilities.  can actually use any .png image as a brush (if saved as brush_X.png and placed in the mipav/brushes user folder... similar to mipav/plugins)
brushes are now stored in a combobox in both the ViewToolBarBuilder.java and ViewJFrameTriImage.java
preferences remembers the last brush used and sets that when new frames are opened.  currently saving a brush does not refresh the paintbrush list, but any time a new image is opened, the paintbrush list is updated.

there is still a small bit of testing and documentation to add to the code
there is no limit on the size of the brush you can create/load
the combobox displays each brush's dimension in text, and will render a small icon if the brush is above 20x20 pixels
the built-in brushes (.gifs) are in the view/icons folder and are named according to their shape/size... for now there are 5 (1, 4, 8, 16, 24 pixel squares)... will add some circular brushes also.
tri-planar needs to be more thoroughly tested

git-svn-id: https://citdcbmipav.cit.nih.gov/repos-pub/mipav@1656 ba61647d-9d00-f842-95cd-605cb4296b96