Found image that would not read in. Fixed the bug that has been there for a long time.
When an image was big endian we switch form little endian (all 0002 tags are little endian) to big endian after reading 
in past the meta group length (0002, 0000). This length indicates how many bytes are in the 0002 group AFTER the tag 0002,0000)
not from the START of the tag. I tested on numberous DICOM images and they all still are read in correctly and so 
is the new image.



git-svn-id: https://citdcbmipav.cit.nih.gov/repos-pub/mipav@1922 ba61647d-9d00-f842-95cd-605cb4296b96