Fixed loading of metadata from some nifti files.
When reading a .nii or .nii.gz file, check to see if there is a matching .json file with metadata in the same directory.  Can be named with the same base name, or with basename + "_defaced".  Added methods to extract the json metadata and add to the extracted fields that populate the FS DEs (copied from BIDS processing).

Added support for not generating/writing preview thumbnail files/fields (for spectroscopy).

When loading data via CSV for a multi-select DE, put any values not in the PVs into the matching OTH DE, separated by semi-colon.

Fixed memory leak when processing large amounts of spectroscopy data.

Formatted file.

git-svn-id: https://citdcbmipav.cit.nih.gov/repos-pub/mipav@16630 ba61647d-9d00-f842-95cd-605cb4296b96